requirements.txt
jsonschema>=2.6.0
PyYAML>=3.12
